+++ src/vybe/util.clj
@@ -2,6 +2,10 @@
   (:require
    [clojure.string :as str]))
 
+(defonce prd?
+  (or (= (System/getenv "VYBE_PROD") "true")
+      (= (System/getProperty "VYBE_PROD") "true")))
+
 (defonce *state (atom {:debug false}))
 #_(defonce *state (atom {:debug true}))
 #_ (swap! *state assoc :debug true)
@@ -12,9 +16,23 @@
   (swap! *state assoc :debug v))
 
 (defmacro debug
+  "Print vybe debug message.
+
+  It can be enabled by calling `debug-set!` only when the env var or jvm prop
+  VYBE_PROD is not set to \"true\", otherwise it just returns `nil`."
   [& strs]
-  `(when (:debug @*state)
-     (println (str "[Vybe] - " (str/join " " [~@strs])))))
+  (when-not prd?
+    `(when (:debug @*state)
+       (println (str "[Vybe] - " (str/join " " [~@strs]))))))
+
+(defmacro if-prd
+  "Runs `prd-body`, it can be enabled by calling `debug-set!`."
+  [prd-body else-body]
+  (if prd?
+    `(do ~prd-body)
+    `(if (:debug @*state)
+       (do ~else-body)
+       (do ~prd-body))))
 
 (defonce *commands (atom []))
 
@@ -23,3 +41,11 @@
   call."
   [f]
   (swap! *commands conj f))
+
+(defonce *probe (atom {}))
+
+(defn counter!
+  "Used for debugging."
+  [k]
+  (swap! *probe update-in [::counter k] (fnil inc 0)))
+#_ (counter! :a)